Updated README.md

+++ bin/newproject.sh
@@ -37,29 +37,37 @@ mv "$BASE_DIR/$projectName/project.gitignore" "$BASE_DIR/$projectName/.gitignore
 > "$BASE_DIR/$projectName/TODO.txt"
 echo -e "#!/usr/bin/env bash\n\ngit pull && git push" > "$BASE_DIR/$projectName/update.sh"
 
-files=('bin'
-	'docs'
-	'lib/Goji'
-	'public/css/goji.css'
-	'public/css/lib/Goji'
-	'public/js/lib/Goji')
+read -p "Project created, share library files with a clean Goji repository ? (y/n): " useSymlinks
 
-for file in "${files[@]}"
-do
-	read -p "Replace '$file' with a symlink? Type 'y' to confirm: " confirm
-
-	if [ $confirm = 'y' ]
-	then
-		if [ -e "$BASE_DIR/$projectName/$file" ]; then
-			rm -rf "$BASE_DIR/$projectName/$file"
-		fi
-
-		ln -sf "$BASE_DIR/goji/$file" "$BASE_DIR/$projectName/$file"
-	else
-		echo "Did not replace '$file' with a symlink."
+if [ $useSymlinks = 'y' ];
+then
+	if [ ! -d "$BASE_DIR/goji" ]; then
+		git clone https://github.com/qrichert/goji.git "$BASE_DIR/goji"
 	fi
-done
+
+	files=('bin'
+		'docs'
+		'lib/Goji'
+		'public/css/goji.css'
+		'public/css/lib/Goji'
+		'public/js/lib/Goji')
+
+	for file in "${files[@]}"
+	do
+		read -p "Replace '$file' with a symlink? Type 'y' to confirm: " confirm
+
+		if [ $confirm = 'y' ]
+		then
+			if [ -e "$BASE_DIR/$projectName/$file" ]; then
+				rm -rf "$BASE_DIR/$projectName/$file"
+			fi
+
+			ln -sf "$BASE_DIR/goji/$file" "$BASE_DIR/$projectName/$file"
+		else
+			echo "Did not replace '$file' with a symlink."
+		fi
+	done
+fi
 
 echo "Done."
 echo "New project created in '$BASE_DIR/$projectName'"
-
